feat: add string validation

+++ internal/handlers/swift_handlers.go
@@ -5,11 +5,13 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+
 	"strings"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/tomasz-trela/remitly-task/internal/models"
 	"github.com/tomasz-trela/remitly-task/internal/repository"
+	validator "github.com/tomasz-trela/remitly-task/internal/validators"
 )
 
 func writeJSONError(w http.ResponseWriter, status int, message string) {
@@ -26,6 +28,11 @@ func writeJSONMessage(w http.ResponseWriter, status int, message string) {
 
 func GetSwiftDataBySwiftCode(w http.ResponseWriter, r *http.Request) {
 	swiftCode := chi.URLParam(r, "swiftCode")
+
+	if !validator.IsValidSwiftCode(swiftCode) {
+		writeJSONError(w, http.StatusBadRequest, "Invalid swift code format")
+		return
+	}
 
 	swiftCodeResponse, err := repository.GetBankCodeAndBranchesBySwift(swiftCode)
 	if err == sql.ErrNoRows {
@@ -44,11 +51,17 @@ func GetSwiftDataBySwiftCode(w http.ResponseWriter, r *http.Request) {
 func GetSwiftDataByCountryISO2(w http.ResponseWriter, r *http.Request) {
 	countryISO2 := chi.URLParam(r, "countryISO2code")
 
+	if !validator.IsValidISO2(countryISO2) {
+		writeJSONError(w, http.StatusBadRequest, "Invalid ISO2 country code format")
+		return
+	}
+
 	banks, err := repository.GetBanksByISO2(countryISO2)
 	if err == sql.ErrNoRows {
 		writeJSONError(w, http.StatusNotFound, "No banks found for the given country ISO2 code")
 		return
 	}
+
 	if err != nil {
 		writeJSONError(w, http.StatusInternalServerError, "Error retrieving bank data")
 		return
@@ -66,14 +79,17 @@ func PostSwift(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if len(swiftCode.CountryISO2) != 2 {
-		writeJSONError(w, http.StatusBadRequest, "countryISO2 should have length of 2")
+	swiftCode.CountryISO2 = strings.ToUpper(swiftCode.CountryISO2)
+	swiftCode.SwiftCode = strings.ToUpper(swiftCode.SwiftCode)
+
+	if !validator.IsValidISO2(swiftCode.CountryISO2) {
+		writeJSONError(w, http.StatusBadRequest, "countryISO2 must be 2 uppercase letters")
+		return
 	}
 
-	swiftCode.CountryISO2 = strings.ToUpper(swiftCode.CountryISO2)
-
-	if len(swiftCode.SwiftCode) != 11 {
-		writeJSONError(w, http.StatusBadRequest, "Swift code should have length of 11")
+	if !validator.IsValidSwiftCode(swiftCode.SwiftCode) {
+		writeJSONError(w, http.StatusBadRequest, "Swift code must be 11 alphanumeric characters")
+		return
 	}
 
 	rowsAffected, err := repository.InsertSwiftCode(&swiftCode)
@@ -92,6 +108,11 @@ func PostSwift(w http.ResponseWriter, r *http.Request) {
 
 func DeleteSwiftCode(w http.ResponseWriter, r *http.Request) {
 	swiftCode := chi.URLParam(r, "swiftCode")
+
+	if !validator.IsValidSwiftCode(swiftCode) {
+		writeJSONError(w, http.StatusBadRequest, "Invalid swift code format")
+		return
+	}
 
 	rowsAffected, err := repository.DeleteSwiftCode(swiftCode)
 	if err != nil {
